LanaApp_V1/Backend/FastAPI/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic-settings==1.0.0
pydantic[email]==1.10.7
alembic==1.11.1